+++ src/fmt/strtime/mod.rs
@@ -162,6 +162,7 @@ strings, the strings are matched without regard to ASCII case.
 | `%A`, `%a` | `Sunday`, `Sun` | The full and abbreviated weekday, respectively. |
 | `%B`, `%b`, `%h` | `June`, `Jun`, `Jun` | The full and abbreviated month name, respectively. |
 | `%C` | `20` | The century of the year. No padding. |
+| `%c` | `2024 M07 14, Sun 17:31:59` | The date and clock time via [`Custom`]. Supported when formatting only. |
 | `%D` | `7/14/24` | Equivalent to `%m/%d/%y`. |
 | `%d`, `%e` | `25`, ` 5` | The day of the month. `%d` is zero-padded, `%e` is space padded. |
 | `%F` | `2024-07-14` | Equivalent to `%Y-%m-%d`. |
@@ -182,6 +183,7 @@ strings, the strings are matched without regard to ASCII case.
 | `%Q` | `America/New_York`, `+0530` | An IANA time zone identifier, or `%z` if one doesn't exist. |
 | `%:Q` | `America/New_York`, `+05:30` | An IANA time zone identifier, or `%:z` if one doesn't exist. |
 | `%R` | `23:30` | Equivalent to `%H:%M`. |
+| `%r` | `8:30:00 AM` | The 12-hour clock time via [`Custom`]. Supported when formatting only. |
 | `%S` | `59` | The second. Zero padded. |
 | `%s` | `1737396540` | A Unix timestamp, in seconds. |
 | `%T` | `23:30:59` | Equivalent to `%H:%M:%S`. |
@@ -191,6 +193,8 @@ strings, the strings are matched without regard to ASCII case.
 | `%V` | `05` | Week number in the [ISO 8601 week-based] calendar. Zero padded. |
 | `%W` | `03` | Week number. Week 1 is the first week starting with a Monday. Zero padded. |
 | `%w` | `0` | The day of the week beginning with Sunday at `0`. |
+| `%X` | `17:31:59` | The clock time via [`Custom`]. Supported when formatting only. |
+| `%x` | `2024 M07 14` | The date via [`Custom`]. Supported when formatting only. |
 | `%Y` | `2024` | A full year, including century. Zero padded to 4 digits. |
 | `%y` | `24` | A two-digit year. Represents only 1969-2068. Zero padded. |
 | `%Z` | `EDT` | A time zone abbreviation. Supported when formatting only. |
@@ -210,8 +214,8 @@ entirely in uppercase by default.
 
 The above flags override the "default" settings of a specifier. For example,
 `%_d` pads with spaces instead of zeros, and `%0e` pads with zeros instead of
-spaces. The exceptions are the `%z` and `%:z` specifiers. They are unaffected
-by any flags.
+spaces. The exceptions are the locale (`%c`, `%r`, `%X`, `%x`), and time zone
+(`%z`, `%:z`) specifiers. They are unaffected by any flags.
 
 Moreover, any number of decimal digits can be inserted after the (possibly
 absent) flag and before the directive, so long as the parsed number is less
@@ -247,9 +251,17 @@ is variable width data. If you have a use case for this, please
 The following things are currently unsupported:
 
 * Parsing or formatting fractional seconds in the time time zone offset.
-* Locale oriented conversion specifiers, such as `%c`, `%r` and `%+`, are not
-  supported by Jiff. For locale oriented datetime formatting, please use the
-  [`icu`] crate via [`jiff-icu`].
+* The `%::z` and `%:::z` specifiers found in GNU date.
+* The `%+` conversion specifier is not supported since there doesn't seem to
+  be any consistent definition for it.
+* With only Jiff, the `%c`, `%r`, `%X` and `%x` locale oriented specifiers
+  use a default "unknown" locale via the [`DefaultCustom`] implementation
+  of the [`Custom`] trait. An example of the default locale format for `%c`
+  is `2024 M07 14, Sun 17:31:59`. One can either switch the POSIX locale
+  via [`PosixCustom`] (e.g., `Sun Jul 14 17:31:59 2024`), or write your own
+  implementation of [`Custom`] powered by [`icu`] and glued together with Jiff
+  via [`jiff-icu`].
+* The `E` and `O` locale modifiers are not supported.
 
 [`strftime`]: https://pubs.opengroup.org/onlinepubs/009695399/functions/strftime.html
 [`strptime`]: https://pubs.opengroup.org/onlinepubs/009695399/functions/strptime.html
@@ -421,6 +433,315 @@ pub fn format(
     let mut buf = alloc::string::String::new();
     broken_down_time.format(format, &mut buf)?;
     Ok(buf)
+}
+
+/// Configuration for customizing the behavior of formatting or parsing.
+///
+/// Currently, this type is limited to being a vehicle for setting the
+/// [`Custom`] trait implementation to use when calling
+/// [`BrokenDownTime::format_with_config`]
+/// or [`BrokenDownTime::to_string_with_config`].
+///
+/// More functionality may be added in the future.
+///
+/// It is generally expected that most callers should not need to use this.
+/// At present, the only reason to use this is if you specifically need to
+/// provide locale aware formatting within the context of `strtime`-style
+/// APIs. Unless you specifically need this, you should prefer using the
+/// [`icu`] crate via [`jiff-icu`] to do type conversions. More specifically,
+/// follow the examples in the `icu::datetime` module for a modern approach
+/// to datetime localization that leverages Unicode.
+///
+/// # Example
+///
+/// This example shows how to use [`PosixCustom`] via `strtime` formatting:
+///
+/// ```
+/// use jiff::{civil, fmt::strtime::{BrokenDownTime, PosixCustom, Config}};
+///
+/// let config = Config::new().custom(PosixCustom::new());
+/// let dt = civil::date(2025, 7, 1).at(17, 30, 0, 0);
+/// let tm = BrokenDownTime::from(dt);
+/// assert_eq!(
+///     tm.to_string_with_config(&config, "%c")?,
+///     "Tue Jul  1 17:30:00 2025",
+/// );
+///
+/// # Ok::<(), Box<dyn std::error::Error>>(())
+/// ```
+///
+/// [`icu`]: https://docs.rs/icu
+/// [`jiff-icu`]: https://docs.rs/jiff-icu
+#[derive(Clone, Debug)]
+pub struct Config<C> {
+    custom: C,
+}
+
+impl Config<DefaultCustom> {
+    /// Create a new default `Config` that uses [`DefaultCustom`].
+    pub const fn new() -> Config<DefaultCustom> {
+        Config { custom: DefaultCustom::new() }
+    }
+}
+
+impl<C> Config<C> {
+    /// Set the implementation of [`Custom`] to use in `strtime`-style APIs
+    /// that use this configuration.
+    pub fn custom<U: Custom>(self, custom: U) -> Config<U> {
+        Config { custom }
+    }
+}
+
+/// An interface for customizing `strtime`-style parsing and formatting.
+///
+/// Each method on this trait comes with a default implementation corresponding
+/// to the behavior of [`DefaultCustom`]. More methods on this trait may be
+/// added in the future.
+///
+/// Implementors of this trait can be attached to a [`Config`] which can then
+/// be passed to [`BrokenDownTime::format_with_config`] or
+/// [`BrokenDownTime::to_string_with_config`].
+///
+/// New methods with default implementations may be added to this trait in
+/// semver compatible releases of Jiff.
+///
+/// # Motivation
+///
+/// While Jiff's API is generally locale-agnostic, this trait is meant to
+/// provide a best effort "hook" for tailoring the behavior of `strtime`
+/// routines. More specifically, for conversion specifiers in `strtime`-style
+/// APIs that are influenced by locale settings.
+///
+/// In general, a `strtime`-style API is not optimal for localization.
+/// It's both too flexible and not expressive enough. As a result, mixing
+/// localization with `strtime`-style APIs is likely not a good idea. However,
+/// this is sometimes required for legacy or convenience reasons, and that's
+/// why Jiff provides this hook.
+///
+/// If you do need to localize datetimes but don't have a requirement to
+/// have it integrate with `strtime`-style APIs, then you should use the
+/// [`icu`] crate via [`jiff-icu`] for type conversions. And then follow the
+/// examples in the `icu::datetime` API for formatting datetimes.
+///
+/// # Supported conversion specifiers
+///
+/// Currently, only formatting for the following specifiers is supported:
+///
+/// * `%c` - Formatting the date and time.
+/// * `%r` - Formatting the 12-hour clock time.
+/// * `%X` - Formatting the clock time.
+/// * `%x` - Formatting the date.
+///
+/// # Unsupported behavior
+///
+/// This trait currently does not support parsing based on locale in any way.
+///
+/// This trait also does not support locale specific behavior for `%a`/`%A`
+/// (day of the week), `%b/`%B` (name of the month) or `%p`/`%P` (AM or PM).
+/// Supporting these is problematic with modern localization APIs, since
+/// modern APIs do not expose options to localize these things independent of
+/// anything else. Instead, they are subsumed most holistically into, e.g.,
+/// "print the long form of a date in the current locale."
+///
+/// Since the motivation for this trait is not really to provide the best way
+/// to localize datetimes, but rather, to facilitate convenience and
+/// inter-operation with legacy systems, it is plausible that the behaviors
+/// listed above could be supported by Jiff. If you need the above behaviors,
+/// please [open a new issue](https://github.com/BurntSushi/jiff/issues/new)
+/// with a proposal.
+///
+/// # Example
+///
+/// This example shows the difference between the default locale and the
+/// POSIX locale:
+///
+/// ```
+/// use jiff::{civil, fmt::strtime::{BrokenDownTime, PosixCustom, Config}};
+///
+/// let dt = civil::date(2025, 7, 1).at(17, 30, 0, 0);
+/// let tm = BrokenDownTime::from(dt);
+/// assert_eq!(
+///     tm.to_string("%c")?,
+///     "2025 M07 1, Tue 17:30:00",
+/// );
+///
+/// let config = Config::new().custom(PosixCustom::new());
+/// assert_eq!(
+///     tm.to_string_with_config(&config, "%c")?,
+///     "Tue Jul  1 17:30:00 2025",
+/// );
+///
+/// # Ok::<(), Box<dyn std::error::Error>>(())
+/// ```
+///
+/// [`icu`]: https://docs.rs/icu
+/// [`jiff-icu`]: https://docs.rs/jiff-icu
+pub trait Custom: Sized {
+    /// Called when formatting a datetime with the `%c` flag.
+    ///
+    /// This defaults to the implementation for [`DefaultCustom`].
+    fn format_datetime<W: Write>(
+        &self,
+        config: &Config<Self>,
+        _ext: &Extension,
+        tm: &BrokenDownTime,
+        wtr: &mut W,
+    ) -> Result<(), Error> {
+        tm.format_with_config(config, "%Y M%m %-d, %a %H:%M:%S", wtr)
+    }
+
+    /// Called when formatting a datetime with the `%x` flag.
+    ///
+    /// This defaults to the implementation for [`DefaultCustom`].
+    fn format_date<W: Write>(
+        &self,
+        config: &Config<Self>,
+        _ext: &Extension,
+        tm: &BrokenDownTime,
+        wtr: &mut W,
+    ) -> Result<(), Error> {
+        // 2025 M04 27
+        tm.format_with_config(config, "%Y M%m %-d", wtr)
+    }
+
+    /// Called when formatting a datetime with the `%X` flag.
+    ///
+    /// This defaults to the implementation for [`DefaultCustom`].
+    fn format_time<W: Write>(
+        &self,
+        config: &Config<Self>,
+        _ext: &Extension,
+        tm: &BrokenDownTime,
+        wtr: &mut W,
+    ) -> Result<(), Error> {
+        tm.format_with_config(config, "%H:%M:%S", wtr)
+    }
+
+    /// Called when formatting a datetime with the `%r` flag.
+    ///
+    /// This defaults to the implementation for [`DefaultCustom`].
+    fn format_12hour_time<W: Write>(
+        &self,
+        config: &Config<Self>,
+        _ext: &Extension,
+        tm: &BrokenDownTime,
+        wtr: &mut W,
+    ) -> Result<(), Error> {
+        tm.format_with_config(config, "%-I:%M:%S %p", wtr)
+    }
+}
+
+/// The default trait implementation of [`Custom`].
+///
+/// Whenever one uses the formatting or parsing routines in this module
+/// without providing a configuration, then this customization is the one
+/// that gets used.
+///
+/// The behavior of the locale formatting of this type is meant to match that
+/// of Unicode's `und` locale.
+///
+/// # Example
+///
+/// This example shows how to explicitly use [`DefaultCustom`] via `strtime`
+/// formatting:
+///
+/// ```
+/// use jiff::{civil, fmt::strtime::{BrokenDownTime, DefaultCustom, Config}};
+///
+/// let config = Config::new().custom(DefaultCustom::new());
+/// let dt = civil::date(2025, 7, 1).at(17, 30, 0, 0);
+/// let tm = BrokenDownTime::from(dt);
+/// assert_eq!(
+///     tm.to_string_with_config(&config, "%c")?,
+///     "2025 M07 1, Tue 17:30:00",
+/// );
+///
+/// # Ok::<(), Box<dyn std::error::Error>>(())
+/// ```
+#[derive(Clone, Debug, Default)]
+pub struct DefaultCustom(());
+
+impl DefaultCustom {
+    /// Create a new instance of this default customization.
+    pub const fn new() -> DefaultCustom {
+        DefaultCustom(())
+    }
+}
+
+impl Custom for DefaultCustom {}
+
+/// A POSIX locale implementation of [`Custom`].
+///
+/// The behavior of the locale formatting of this type is meant to match that
+/// of POSIX's `C` locale.
+///
+/// # Example
+///
+/// This example shows how to use [`PosixCustom`] via `strtime` formatting:
+///
+/// ```
+/// use jiff::{civil, fmt::strtime::{BrokenDownTime, PosixCustom, Config}};
+///
+/// let config = Config::new().custom(PosixCustom::new());
+/// let dt = civil::date(2025, 7, 1).at(17, 30, 0, 0);
+/// let tm = BrokenDownTime::from(dt);
+/// assert_eq!(
+///     tm.to_string_with_config(&config, "%c")?,
+///     "Tue Jul  1 17:30:00 2025",
+/// );
+///
+/// # Ok::<(), Box<dyn std::error::Error>>(())
+/// ```
+#[derive(Clone, Debug, Default)]
+pub struct PosixCustom(());
+
+impl PosixCustom {
+    /// Create a new instance of this POSIX customization.
+    pub const fn new() -> PosixCustom {
+        PosixCustom(())
+    }
+}
+
+impl Custom for PosixCustom {
+    fn format_datetime<W: Write>(
+        &self,
+        config: &Config<Self>,
+        _ext: &Extension,
+        tm: &BrokenDownTime,
+        wtr: &mut W,
+    ) -> Result<(), Error> {
+        tm.format_with_config(config, "%a %b %e %H:%M:%S %Y", wtr)
+    }
+
+    fn format_date<W: Write>(
+        &self,
+        config: &Config<Self>,
+        _ext: &Extension,
+        tm: &BrokenDownTime,
+        wtr: &mut W,
+    ) -> Result<(), Error> {
+        tm.format_with_config(config, "%m/%d/%y", wtr)
+    }
+
+    fn format_time<W: Write>(
+        &self,
+        config: &Config<Self>,
+        _ext: &Extension,
+        tm: &BrokenDownTime,
+        wtr: &mut W,
+    ) -> Result<(), Error> {
+        tm.format_with_config(config, "%H:%M:%S", wtr)
+    }
+
+    fn format_12hour_time<W: Write>(
+        &self,
+        config: &Config<Self>,
+        _ext: &Extension,
+        tm: &BrokenDownTime,
+        wtr: &mut W,
+    ) -> Result<(), Error> {
+        tm.format_with_config(config, "%I:%M:%S %p", wtr)
+    }
 }
 
 /// The "broken down time" used by parsing and formatting.
@@ -694,8 +1015,59 @@ impl BrokenDownTime {
         format: impl AsRef<[u8]>,
         mut wtr: W,
     ) -> Result<(), Error> {
+        self.format_with_config(&Config::new(), format, &mut wtr)
+    }
+
+    /// Format this broken down time with a specific configuration using the
+    /// format string given.
+    ///
+    /// See the [module documentation](self) for details on what's supported.
+    ///
+    /// This routine is like [`BrokenDownTime::format`], except that it
+    /// permits callers to provide their own configuration instead of using
+    /// the default. This routine also accepts a `&mut W` instead of a `W`,
+    /// which may be more flexible in some situations.
+    ///
+    /// # Errors
+    ///
+    /// This returns an error when formatting failed. Formatting can fail
+    /// either because of an invalid format string, or if formatting requires
+    /// a field in `BrokenDownTime` to be set that isn't. For example, trying
+    /// to format a [`DateTime`] with the `%z` specifier will fail because a
+    /// `DateTime` has no time zone or offset information associated with it.
+    ///
+    /// Formatting also fails if writing to the given writer fails.
+    ///
+    /// # Example
+    ///
+    /// This example shows how to use [`PosixCustom`] to get formatting
+    /// for conversion specifiers like `%c` in the POSIX locale:
+    ///
+    /// ```
+    /// use jiff::{civil, fmt::strtime::{BrokenDownTime, PosixCustom, Config}};
+    ///
+    /// let mut buf = String::new();
+    /// let dt = civil::date(2025, 7, 1).at(17, 30, 0, 0);
+    /// let tm = BrokenDownTime::from(dt);
+    /// tm.format("%c", &mut buf)?;
+    /// assert_eq!(buf, "2025 M07 1, Tue 17:30:00");
+    ///
+    /// let config = Config::new().custom(PosixCustom::new());
+    /// buf.clear();
+    /// tm.format_with_config(&config, "%c", &mut buf)?;
+    /// assert_eq!(buf, "Tue Jul  1 17:30:00 2025");
+    ///
+    /// # Ok::<(), Box<dyn std::error::Error>>(())
+    /// ```
+    #[inline]
+    pub fn format_with_config<W: Write, L: Custom>(
+        &self,
+        config: &Config<L>,
+        format: impl AsRef<[u8]>,
+        wtr: &mut W,
+    ) -> Result<(), Error> {
         let fmt = format.as_ref();
-        let mut formatter = Formatter { fmt, tm: self, wtr: &mut wtr };
+        let mut formatter = Formatter { config, fmt, tm: self, wtr };
         formatter.format().context("strftime formatting failed")?;
         Ok(())
     }
@@ -743,6 +1115,58 @@ impl BrokenDownTime {
     ) -> Result<alloc::string::String, Error> {
         let mut buf = alloc::string::String::new();
         self.format(format, &mut buf)?;
+        Ok(buf)
+    }
+
+    /// Format this broken down time with a specific configuration using the
+    /// format string given into a new `String`.
+    ///
+    /// See the [module documentation](self) for details on what's supported.
+    ///
+    /// This routine is like [`BrokenDownTime::to_string`], except that it
+    /// permits callers to provide their own configuration instead of using
+    /// the default.
+    ///
+    /// # Errors
+    ///
+    /// This returns an error when formatting failed. Formatting can fail
+    /// either because of an invalid format string, or if formatting requires
+    /// a field in `BrokenDownTime` to be set that isn't. For example, trying
+    /// to format a [`DateTime`] with the `%z` specifier will fail because a
+    /// `DateTime` has no time zone or offset information associated with it.
+    ///
+    /// # Example
+    ///
+    /// This example shows how to use [`PosixCustom`] to get formatting
+    /// for conversion specifiers like `%c` in the POSIX locale:
+    ///
+    /// ```
+    /// use jiff::{civil, fmt::strtime::{BrokenDownTime, PosixCustom, Config}};
+    ///
+    /// let dt = civil::date(2025, 7, 1).at(17, 30, 0, 0);
+    /// let tm = BrokenDownTime::from(dt);
+    /// assert_eq!(
+    ///     tm.to_string("%c")?,
+    ///     "2025 M07 1, Tue 17:30:00",
+    /// );
+    ///
+    /// let config = Config::new().custom(PosixCustom::new());
+    /// assert_eq!(
+    ///     tm.to_string_with_config(&config, "%c")?,
+    ///     "Tue Jul  1 17:30:00 2025",
+    /// );
+    ///
+    /// # Ok::<(), Box<dyn std::error::Error>>(())
+    /// ```
+    #[cfg(feature = "alloc")]
+    #[inline]
+    pub fn to_string_with_config<L: Custom>(
+        &self,
+        config: &Config<L>,
+        format: impl AsRef<[u8]>,
+    ) -> Result<alloc::string::String, Error> {
+        let mut buf = alloc::string::String::new();
+        self.format_with_config(config, format, &mut buf)?;
         Ok(buf)
     }
 
@@ -2643,10 +3067,15 @@ impl From<Time> for Meridiem {
 
 /// These are "extensions" to the standard `strftime` conversion specifiers.
 ///
-/// Basically, these provide control over padding (zeros, spaces or none),
-/// how much to pad and the case of string enumerations.
-#[derive(Clone, Copy, Debug)]
-struct Extension {
+/// This type represents which flags and/or padding were provided with a
+/// specifier. For example, `%_3d` uses 3 spaces of padding.
+///
+/// Currently, this type provides no structured introspection facilities. It
+/// is exported and available only via implementations of the [`Custom`] trait
+/// for reasons of semver compatible API evolution. If you have use cases for
+/// introspecting this type, please open an issue.
+#[derive(Clone, Debug)]
+pub struct Extension {
     flag: Option<Flag>,
     width: Option<u8>,
 }